// src/modules/github-event-manager/achievibit-event-names.enum.ts
export enum AchievibitEventNames {
  NewConnection = 'NewConnection',
  PullRequestOpened = 'PullRequestOpened',
  PullRequestInitialLabeled = 'PullRequestInitialLabeled',
  PullRequestLableAdded = 'PullRequestLableAdded',
  PullRequestLabelRemoved = 'PullRequestLabelRemoved',
  PullRequestEdited = 'PullRequestEdited',
  PullRequestAssigneeAdded = 'PullRequestAssigneeAdded',
  PullRequestAssigneeRemoved = 'PullRequestAssigneeRemoved',
  PullRequestReviewRequestAdded = 'PullRequestReviewRequestAdded',
  PullRequestReviewRequestRemoved = 'PullRequestReviewRequestRemoved',
  PullRequestReviewCommentAdded = 'PullRequestReviewCommentAdded',
  PullRequestReviewCommentRemoved = 'PullRequestReviewCommentRemoved',
  PullRequestReviewCommentEdited = 'PullRequestReviewCommentEdited',
  PullRequestReviewSubmitted = 'PullRequestReviewSubmitted',
  PullRequestMerged = 'PullRequestMerged'
}